requirements.txt
torch
diffusers
transformers
Pillow
numpy
